hooks: Overhaul to use GOPROXY
Using GOPROXY to fetch dependencies is more robust than vendoring which requires a patched Go to allow opting out of integrity checks.
This patch is carried in nixpkgs, but with the change in approach we can work with vanilla Go too, and potentially allow nixpkgs to drop this patch in the future.

The vendoring approach is what gomod2nix uses as it predates GOPROXY.

This allows us to work more with the Go toolchain as intended rather than against it.